new figs

+++ notebooks/figs/tuned_boosting.py
@@ -23,7 +23,7 @@ DATASETS_CLASSIFICATION = [
     # ("breast-cancer", "breast_cancer", 'imodels'), # this is the wrong breast-cancer dataset (https://new.openml.org/search?type=data&sort=runs&id=13&status=active)
     # ("haberman", "haberman", 'imodels'),
     # ("ionosphere", "ionosphere", 'pmlb'),
-    # ("diabetes", "diabetes", "pmlb"),
+    ("diabetes", "diabetes", "pmlb"),
     # ("liver", "8", "openml"), # note: we omit this dataset bc it's label was found to be incorrect (see caveat here: https://archive.ics.uci.edu/ml/datasets/liver+disorders#:~:text=The%207th%20field%20(selector)%20has%20been%20widely%20misinterpreted%20in%20the%20past%20as%20a%20dependent%20variable%20representing%20presence%20or%20absence%20of%20a%20liver%20disorder.)
     # ("credit-g", "credit_g", 'imodels'), # like german-credit, but more feats
     ("german-credit", "german", "pmlb"),
@@ -134,7 +134,7 @@ def figs_vs_boosting(X, y, budget,depth, n_seeds=10, only_boosting=False):
             gb_model = GradientBoostingClassifier if is_classification else GradientBoostingRegressor
             gb = gb_model(n_estimators=n_estimators, max_depth=depth)
             gb.fit(X_train, y_train)
-            preds = gb.predict_proba(X_test) if is_classification else gb.predict(X_test)
+            preds = gb.predict_proba(X_test)[:, 1] if is_classification else gb.predict(X_test)
             gb_score = metric(y_test, preds)
             scores["boosting"].append(gb_score)
         else:
@@ -146,7 +146,7 @@ def figs_vs_boosting(X, y, budget,depth, n_seeds=10, only_boosting=False):
         figs_model = FIGSClassifier if is_classification else FIGSRegressor
         figs = figs_model(max_rules=budget)
         figs.fit(X_train, y_train)
-        preds_figs = figs.predict_proba(X_test) if is_classification else figs.predict(X_test)
+        preds_figs = figs.predict_proba(X_test)[:, 1] if is_classification else figs.predict(X_test)
         figs_score = metric(y_test, preds_figs)
         scores["figs"].append(figs_score)
 
@@ -163,7 +163,7 @@ def analyze_datasets(datasets, fig_name=None):
     # make a list of multipication of n_rules_per_tree as long as it is less than 20
     budgets = np.arange(5, 21)
 
-    fig, axes = plt.subplots(n_rows, n_cols, figsize=(15, 15))
+    fig, axes = plt.subplots(n_rows, n_cols, figsize=(15, 10))
     n_seeds = 20
     for i, d in enumerate(datasets):
         if isinstance(d, str):
@@ -172,47 +172,57 @@ def analyze_datasets(datasets, fig_name=None):
             dset_name = d[0]
         ax = axes[i // n_cols, i % n_cols]
         X, y, feat_names = get_clean_dataset(d[1], data_source=d[2])
-        means = {"figs": [], "boosting d1": [], "boosting d2": [], "boosting d3":[]}
-        std = {"figs": [], "boosting d1": [], "boosting d2": [], "boosting d3":[]}
-        for budget in budgets:
-            scores = figs_vs_boosting(X, y, budget=budget, n_seeds=n_seeds, depth=1)
-            means["figs"].append(np.mean(scores["figs"]))
-            means["boosting d1"].append(np.mean(scores["boosting"]))
-            std["figs"].append(np.std(scores["figs"]) / np.sqrt(n_seeds))
-            std["boosting d1"].append(np.std(scores["boosting"]) / np.sqrt(n_seeds))
-            for d in [2,3]:
-                scores = figs_vs_boosting(X, y, budget=budget, n_seeds=n_seeds, depth=d)
-                is_na = np.isnan(scores["boosting"]).sum() > 0
-                if is_na:
-                    # set mean and std to nan
-                    means[f"boosting d{d}"].append(np.nan)
-                    std[f"boosting d{d}"].append(np.nan)
-                    continue
-                means[f"boosting d{d}"].append(np.nanmean(scores["boosting"]))
-                std[f"boosting d{d}"].append(np.nanstd(scores["boosting"]) / np.sqrt(n_seeds))
-        # make plot with error bars vs budget
-        ds_data = {"means": means, "std": std, "budgets": budgets}
-        # save the pickle file
-        fname = f"figs_vs_boosting_{dset_name}_cls.pkl" if len(np.unique(y)) == 2 else f"figs_vs_boosting_{dset_name}_reg.pkl"
-        with open(fname, "wb") as f:
-            pickle.dump(ds_data, f)
+        f_name = f"figs_vs_boosting_{dset_name}_cls.pkl" if len(np.unique(y)) == 2 else f"figs_vs_boosting_{dset_name}_reg.pkl"
+        if os.path.exists(f_name):
+            ds_data = pickle.load(open(f_name, "rb"))
+            means = ds_data["means"]
+            std = ds_data["std"]
+        else:
+            means = {"figs": [], "boosting d1": [], "boosting d2": [], "boosting d3":[]}
+            std = {"figs": [], "boosting d1": [], "boosting d2": [], "boosting d3":[]}
+            for budget in budgets:
+                scores = figs_vs_boosting(X, y, budget=budget, n_seeds=n_seeds, depth=1)
+                means["figs"].append(np.mean(scores["figs"]))
+                means["boosting d1"].append(np.mean(scores["boosting"]))
+                std["figs"].append(np.std(scores["figs"]) / np.sqrt(n_seeds))
+                std["boosting d1"].append(np.std(scores["boosting"]) / np.sqrt(n_seeds))
+                for d in [2,3]:
+                    scores = figs_vs_boosting(X, y, budget=budget, n_seeds=n_seeds, depth=d)
+                    is_na = np.isnan(scores["boosting"]).sum() > 0
+                    if is_na:
+                        # set mean and std to nan
+                        means[f"boosting d{d}"].append(np.nan)
+                        std[f"boosting d{d}"].append(np.nan)
+                        continue
+                    means[f"boosting d{d}"].append(np.nanmean(scores["boosting"]))
+                    std[f"boosting d{d}"].append(np.nanstd(scores["boosting"]) / np.sqrt(n_seeds))
+            # make plot with error bars vs budget
+            ds_data = {"means": means, "std": std, "budgets": budgets}
+            # save the pickle file
+            fname = f"figs_vs_boosting_{dset_name}_cls.pkl" if len(np.unique(y)) == 2 else f"figs_vs_boosting_{dset_name}_reg.pkl"
+            with open(fname, "wb") as f:
+                pickle.dump(ds_data, f)
         ax.errorbar(budgets, means["figs"], yerr=std["figs"], label="FIGS", color=COLORS["FIGS"])
         for depth in [1,2,3]:
-            ax.errorbar(budgets, means[f"boosting d{depth}"], yerr=std[f"boosting d{depth}"], label=f"GB (max_depth = {depth})", color=COLORS[f"GBDT-{depth}"])
+            ax.errorbar(budgets, means[f"boosting d{depth}"], yerr=std[f"boosting d{depth}"], label=f"GB (max_depth = {depth})", color=COLORS[f"GBDT-{depth}"],
+                        alpha=0.5)
         ax.set_title(dset_name.capitalize().replace('-', ' ') + f' ($n={DSET_METADATA.get(dset_name, (-1))[0]}$)',
                   fontsize='medium')
         ax.set_xlabel("# of rules")
         ylab = "AUC" if len(np.unique(y)) == 2 else "R2"
-        ax.set_ylabel(ylab)
-        ax.legend()
+        if i % n_cols == 0:
+            ax.set_ylabel(ylab)
+        if i == 0:
+            ax.legend()
+        # ax.legend()
     plt.tight_layout()
     # fig_name = "figs_vs_boosting_regression.png" if
-    plt.savefig(f"{fig_name}.png")
+    plt.savefig(f"{fig_name}.png", dpi=300)
 
 
 def main():
     # for depth in [1,2,3]:
-    analyze_datasets(DATASETS_CLASSIFICATION, fig_name=f"figs_vs_boosting_classification")
+    analyze_datasets(DATASETS_CLASSIFICATION, fig_name=f"figs_vs_boosting_classification_new")
     analyze_datasets(DATASETS_REGRESSION, fig_name=f"figs_vs_boosting_regression")
